+++ src/dotnet/APIView/APIViewWeb/LeanControllers/PermissionsCookieOrTokenController.cs
@@ -1,7 +1,8 @@
+using System;
 using System.Collections.Generic;
+using System.Linq;
 using System.Threading.Tasks;
 using APIViewWeb.Helpers;
-using APIViewWeb.LeanModels;
 using APIViewWeb.Managers.Interfaces;
 using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Http;
@@ -21,13 +22,29 @@ public class PermissionsCookieOrTokenController : ControllerBase
         _permissionsManager = permissionsManager;
     }
 
+
     /// <summary>
-    ///     Get all permission groups
+    ///     Get the list of approvers for a specific language
     /// </summary>
-    [HttpGet("groups")]
-    public async Task<ActionResult> GetAllGroups()
+    /// <param name="language">The programming language</param>
+    /// <returns>List of usernames who can approve reviews for the specified language, sorted alphabetically</returns>
+    [HttpGet("approvers/{language}")]
+    public async Task<ActionResult<IEnumerable<string>>> GetApproversForLanguage(string language)
     {
-        IEnumerable<GroupPermissionsModel> groups = await _permissionsManager.GetAllGroupsAsync();
-        return new LeanJsonResult(groups, StatusCodes.Status200OK);
+        HashSet<string> approvers = await _permissionsManager.GetApproversForLanguageAsync(language);
+        List<string> sortedApprovers = approvers.Where(a => !string.IsNullOrWhiteSpace(a))
+            .OrderBy(a => a, StringComparer.OrdinalIgnoreCase).ToList();
+        return new LeanJsonResult(sortedApprovers, StatusCodes.Status200OK);
+    }
+
+    /// <summary>
+    ///     Get the list of admin usernames for contact information
+    /// </summary>
+    /// <returns>List of usernames who have admin permissions</returns>
+    [HttpGet("admins")]
+    public async Task<ActionResult<IEnumerable<string>>> GetAdminUsernames()
+    {
+        IEnumerable<string> admins = await _permissionsManager.GetAdminUsernamesAsync();
+        return new LeanJsonResult(admins, StatusCodes.Status200OK);
     }
 }
